:rocket: :fire: Timeline Vec to HashMap, change primitives
The timeline original Vec got swapped out for a HashMap. I might go back
and change it back to a pair of `Vec`s. The performance boost from this
will be minimal, but for bigger games it might make a change.

The timeline coordinate, which was originally a f32, is now a i32.
Again, no crazy performance boost out of this, but still one
nonetheless.

The parsing of games is now done using serde and serde_json. This means
less verbosity in the errors, though.

width, height, x and y got swapped out for a u8; this is the reason as
to why almost every file got changed.